게시물 삭제 권한 부여

+++ src/main/java/com/KMS/exam/demo/controller/UsrArticleController.java
@@ -3,7 +3,6 @@ package com.KMS.exam.demo.controller;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,8 +10,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.KMS.exam.demo.service.ArticleService;
+import com.KMS.exam.demo.service.MemberService;
 import com.KMS.exam.demo.util.Ut;
 import com.KMS.exam.demo.vo.Article;
+import com.KMS.exam.demo.vo.Member;
 import com.KMS.exam.demo.vo.ResultData;
 
 
@@ -22,17 +23,17 @@ public class UsrArticleController {
 	// 인스턴스 변수
 	@Autowired
 	private ArticleService articleService;
-
+	@Autowired
+	private MemberService memberService;
+	
 	// 액션메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
 	public ResultData<Article> doAdd(String title, String body ,HttpServletRequest request) {
 		
-		HttpSession session = request.getSession();
+		int loginedId = Ut.getLoginedId(request);
 		
-		int loginedId = (int) session.getAttribute("loginedId");
-		
-		if(session == null || loginedId == 0){
+		if(loginedId == 0){
 			return ResultData.from("F-1", "로그인 하세요");
 		}
 		if(Ut.empty(title)) {
@@ -61,15 +62,26 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Article> doDelete(int id) {
+	public ResultData<Article> doDelete(int id, HttpServletRequest request) {
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
+		int loginedId = Ut.getLoginedId(request);
+		
+		if(loginedId == 0) {
+			return ResultData.from("F-2", Ut.f("로그인 하지 않았습니다."));
+		}
+		
+		Member member = memberService.getMember(loginedId);
+		int loginedLevel = member.getAuthLevel();
+		
+		if(loginedId != article.getLoginedId() &&  loginedLevel != 7) {
+			return ResultData.from("F-3", Ut.f("권한이 없습니다."));
+		}
 
 		articleService.deleteArticle(id);
-
 		return ResultData.from("S-1", Ut.f("%d번 게시물삭제.", id), article);
 	}
